Rename ParserPuppet -> InteractiveParser
